// src/generated/models/Ai_alhasbaincidentsesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Ai_alhasbaincidentsesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Ai_alhasbaincidentsesstatecode = keyof typeof Ai_alhasbaincidentsesstatecode;
export const Ai_alhasbaincidentsesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Ai_alhasbaincidentsesstatuscode = keyof typeof Ai_alhasbaincidentsesstatuscode;

export interface Ai_alhasbaincidentsesBase {
  ai_agent_id?: string;
  ai_alhasbaincidentsid: string;
  ai_change_management_triggered?: boolean;
  ai_comments?: string;
  ai_description?: string;
  ai_division?: string;
  ai_id?: string;
  ai_reported_date?: string;
  ai_resolution?: string;
  ai_resolved_date?: string;
  ai_severity?: string;
  ai_status?: string;
  ai_submitter?: string;
  ai_title: string;
  ai_type?: string;
  ai_use_case_id?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Ai_alhasbaincidentsesstatecode;
  statuscode?: Ai_alhasbaincidentsesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Ai_alhasbaincidentses extends Ai_alhasbaincidentsesBase {
  ai_change_management_triggeredname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}